segment test classes and behaviors
