+++ Packages/mygamedevtools-scene-loader/Tests/Runtime/SceneLoaderTests.cs
@@ -1,8 +1,6 @@
 #if ENABLE_UNITASK
-using Cysharp.Threading.Tasks;
 #endif
 using NUnit.Framework;
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -21,11 +19,11 @@ namespace MyGameDevTools.SceneLoading.Tests
         Coroutine,
     }
 
-    public class SceneLoaderTests
+    public class SceneLoaderTests : SceneTestBase
     {
-        const int _defaultTimeout = 3000;
-
-        static readonly ILoadSceneInfo[] _loadingSceneInfos = new ILoadSceneInfo[]
+        // Here we don't need to test multiple load scene infos, since that is already done
+        // in the scene manager tests.
+        public static readonly ILoadSceneInfo[] LoadingSceneInfos = new ILoadSceneInfo[]
         {
             null,
             new LoadSceneInfoName(SceneBuilder.SceneNames[3]),
@@ -41,26 +39,6 @@ namespace MyGameDevTools.SceneLoading.Tests
             new SceneLoaderCoroutine(SceneTestEnvironment.SceneManagers[0]),
         };
 
-        // Note: These functions must create new scene managers to correctly test the dispose flow
-        static readonly Func<ISceneLoader>[] _sceneLoaderCreateFuncs = new Func<ISceneLoader>[]
-        {
-            () => new SceneLoaderAsync(new AdvancedSceneManager()),
-#if ENABLE_UNITASK
-            () => new SceneLoaderUniTask(new AdvancedSceneManager()),
-#endif
-            () => new SceneLoaderCoroutine(new AdvancedSceneManager()),
-        };
-
-        [UnityTearDown]
-        public IEnumerator TearDown()
-        {
-            foreach (var m in SceneTestEnvironment.SceneManagers)
-                yield return SceneTestUtilities.UnloadManagerScenes(m);
-
-            yield return SceneTestUtilities.UnloadRemainingScenes();
-            Assert.AreEqual(1, SceneManager.loadedSceneCount);
-        }
-
         [UnityTest]
         public IEnumerator LoadScenes([ValueSource(nameof(_sceneLoaders))] ISceneLoader sceneLoader, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes)
         {
@@ -72,7 +50,7 @@ namespace MyGameDevTools.SceneLoading.Tests
 
             var watch = new Stopwatch();
             watch.Start();
-            yield return new WaitUntil(() => loadedScenes.Count == sceneCount || watch.ElapsedMilliseconds > _defaultTimeout);
+            yield return new WaitUntil(() => loadedScenes.Count == sceneCount || watch.ElapsedMilliseconds > SceneTestEnvironment.DefaultTimeout);
             watch.Stop();
 
             sceneLoader.Manager.SceneLoaded -= sceneLoaded;
@@ -99,7 +77,7 @@ namespace MyGameDevTools.SceneLoading.Tests
 
             var watch = new Stopwatch();
             watch.Start();
-            yield return new WaitUntil(() => sceneLoader.Manager.LoadedSceneCount == sceneCount || watch.ElapsedMilliseconds > _defaultTimeout);
+            yield return new WaitUntil(() => sceneLoader.Manager.LoadedSceneCount == sceneCount || watch.ElapsedMilliseconds > SceneTestEnvironment.DefaultTimeout);
             watch.Stop();
 
             Assert.AreEqual(sceneCount, sceneLoader.Manager.LoadedSceneCount);
@@ -110,7 +88,7 @@ namespace MyGameDevTools.SceneLoading.Tests
             sceneLoader.UnloadScenes(targetScenes);
 
             watch.Restart();
-            yield return new WaitUntil(() => unloadedScenes.Count == sceneCount || watch.ElapsedMilliseconds > _defaultTimeout);
+            yield return new WaitUntil(() => unloadedScenes.Count == sceneCount || watch.ElapsedMilliseconds > SceneTestEnvironment.DefaultTimeout);
             watch.Stop();
 
             sceneLoader.Manager.SceneUnloaded -= sceneUnloaded;
@@ -138,12 +116,12 @@ namespace MyGameDevTools.SceneLoading.Tests
 
             var watch = new Stopwatch();
             watch.Start();
-            yield return new WaitUntil(() => unloadedScene.handle != 0 && !unloadedScene.isLoaded || watch.ElapsedMilliseconds > _defaultTimeout);
+            yield return new WaitUntil(() => unloadedScene.handle != 0 && !unloadedScene.isLoaded || watch.ElapsedMilliseconds > SceneTestEnvironment.DefaultTimeout);
             watch.Stop();
 
             sceneLoader.Manager.SceneUnloaded -= sceneUnloaded;
 
-            Assert.Less(watch.ElapsedMilliseconds, _defaultTimeout);
+            Assert.Less(watch.ElapsedMilliseconds, SceneTestEnvironment.DefaultTimeout);
             Assert.AreEqual(loadedScene, unloadedScene);
             Assert.IsFalse(unloadedScene.isLoaded);
 
@@ -154,7 +132,7 @@ namespace MyGameDevTools.SceneLoading.Tests
         }
 
         [UnityTest]
-        public IEnumerator TransitionToScenes([ValueSource(nameof(_sceneLoaders))] ISceneLoader sceneLoader, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes, [ValueSource(nameof(_loadingSceneInfos))] ILoadSceneInfo loadingScene)
+        public IEnumerator TransitionToScenes([ValueSource(nameof(_sceneLoaders))] ISceneLoader sceneLoader, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes, [ValueSource(nameof(LoadingSceneInfos))] ILoadSceneInfo loadingScene)
         {
             yield return LoadFirstScene(sceneLoader);
 
@@ -169,7 +147,7 @@ namespace MyGameDevTools.SceneLoading.Tests
 
             var watch = new Stopwatch();
             watch.Start();
-            yield return new WaitUntil(() => loadedScenes.Count == sceneCount || watch.ElapsedMilliseconds > _defaultTimeout);
+            yield return new WaitUntil(() => loadedScenes.Count == sceneCount || watch.ElapsedMilliseconds > SceneTestEnvironment.DefaultTimeout);
             watch.Stop();
 
             sceneLoader.Manager.SceneLoaded -= sceneLoaded;
@@ -185,7 +163,7 @@ namespace MyGameDevTools.SceneLoading.Tests
         }
 
         [UnityTest]
-        public IEnumerator Transition([ValueSource(nameof(_sceneLoaders))] ISceneLoader sceneLoader, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo targetScene, [ValueSource(nameof(_loadingSceneInfos))] ILoadSceneInfo loadingScene)
+        public IEnumerator Transition([ValueSource(nameof(_sceneLoaders))] ISceneLoader sceneLoader, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo targetScene, [ValueSource(nameof(LoadingSceneInfos))] ILoadSceneInfo loadingScene)
         {
             yield return LoadFirstScene(sceneLoader);
 
@@ -197,7 +175,7 @@ namespace MyGameDevTools.SceneLoading.Tests
 
             var watch = new Stopwatch();
             watch.Start();
-            yield return new WaitUntil(() => loadedScenes.Count == expectedScenes || watch.ElapsedMilliseconds > _defaultTimeout);
+            yield return new WaitUntil(() => loadedScenes.Count == expectedScenes || watch.ElapsedMilliseconds > SceneTestEnvironment.DefaultTimeout);
             watch.Stop();
 
             sceneLoader.Manager.SceneLoaded -= sceneLoaded;
@@ -213,7 +191,7 @@ namespace MyGameDevTools.SceneLoading.Tests
         }
 
         [UnityTest]
-        public IEnumerator Transition_FromExternalOrigin([ValueSource(nameof(_sceneLoaders))] ISceneLoader sceneLoader, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo targetScene, [ValueSource(nameof(_loadingSceneInfos))] ILoadSceneInfo loadingScene)
+        public IEnumerator Transition_FromExternalOrigin([ValueSource(nameof(_sceneLoaders))] ISceneLoader sceneLoader, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo targetScene, [ValueSource(nameof(LoadingSceneInfos))] ILoadSceneInfo loadingScene)
         {
             yield return SceneManager.LoadSceneAsync(SceneBuilder.SceneNames[1], LoadSceneMode.Additive);
             var currentScene = SceneManager.GetSceneByName(SceneBuilder.SceneNames[1]);
@@ -226,7 +204,7 @@ namespace MyGameDevTools.SceneLoading.Tests
 
             var watch = new Stopwatch();
             watch.Start();
-            yield return new WaitUntil(() => loadedScenes.Count == expectedScenes || watch.ElapsedMilliseconds > _defaultTimeout);
+            yield return new WaitUntil(() => loadedScenes.Count == expectedScenes || watch.ElapsedMilliseconds > SceneTestEnvironment.DefaultTimeout);
             watch.Stop();
 
             sceneLoader.Manager.SceneLoaded -= sceneLoaded;
@@ -241,281 +219,10 @@ namespace MyGameDevTools.SceneLoading.Tests
             }
         }
 
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_Simple([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc)
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            Assert.DoesNotThrow(loader.Dispose);
-            yield return null;
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringLoadScene([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc)
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            loader.LoadScene(new LoadSceneInfoName(SceneBuilder.SceneNames[1]));
-            Assert.DoesNotThrow(loader.Dispose);
-            yield return null;
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringLoadScenes([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes)
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            loader.LoadScenes(targetScenes);
-            Assert.DoesNotThrow(loader.Dispose);
-            yield return null;
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringUnloadScene([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc)
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            yield return LoadFirstScene(loader);
-
-            loader.UnloadScene(new LoadSceneInfoScene(loader.Manager.GetLastLoadedScene()));
-            Assert.DoesNotThrow(loader.Dispose);
-            yield return null;
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringUnloadScenes([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes)
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            int sceneCount = targetScenes.Length;
-            loader.LoadScenes(targetScenes);
-
-            var watch = new Stopwatch();
-            watch.Start();
-            yield return new WaitUntil(() => loader.Manager.LoadedSceneCount == sceneCount || watch.ElapsedMilliseconds > _defaultTimeout);
-            watch.Stop();
-
-            loader.UnloadScenes(targetScenes);
-            Assert.DoesNotThrow(loader.Dispose);
-            yield return null;
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringTransitionToScene([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo targetScene, [ValueSource(nameof(_loadingSceneInfos))] ILoadSceneInfo loadingScene)
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            yield return LoadFirstScene(loader);
-
-            loader.TransitionToScene(targetScene, loadingScene);
-            Assert.DoesNotThrow(loader.Dispose);
-            yield return null;
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringTransitionToScenes([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes, [ValueSource(nameof(_loadingSceneInfos))] ILoadSceneInfo loadingScene)
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            yield return LoadFirstScene(loader);
-
-            loader.TransitionToScenes(targetScenes, 0, loadingScene);
-            Assert.DoesNotThrow(loader.Dispose);
-            yield return null;
-        }
-
-#if ENABLE_UNITASK
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringLoadSceneAsync([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo targetScene) => UniTask.ToCoroutine(async () =>
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            SceneLoaderType type = GetSceneLoaderType(loader);
-
-            UniTask<Scene> task = default;
-            task = type switch
-            {
-                SceneLoaderType.Async => ((ISceneLoaderAsync)loader).LoadSceneAsync(targetScene).AsTask().AsUniTask(),
-                SceneLoaderType.UniTask => ((ISceneLoaderUniTask)loader).LoadSceneAsync(targetScene),
-                SceneLoaderType.Coroutine => ((ISceneLoaderCoroutine)loader).LoadSceneAsync(targetScene).Task.AsUniTask(),
-                _ => throw new NotImplementedException($"Type {type} was not implemented"),
-            };
-            loader.Dispose();
-
-            bool canceled = false;
-            try
-            {
-                await task;
-            }
-            catch (OperationCanceledException)
-            {
-                canceled = true;
-            }
-            Assert.True(canceled);
-        });
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringLoadScenesAsync([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes) => UniTask.ToCoroutine(async () =>
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            SceneLoaderType type = GetSceneLoaderType(loader);
-
-            UniTask<Scene[]> task = default;
-            task = type switch
-            {
-                SceneLoaderType.Async => ((ISceneLoaderAsync)loader).LoadScenesAsync(targetScenes).AsTask().AsUniTask(),
-                SceneLoaderType.UniTask => ((ISceneLoaderUniTask)loader).LoadScenesAsync(targetScenes),
-                SceneLoaderType.Coroutine => ((ISceneLoaderCoroutine)loader).LoadScenesAsync(targetScenes).Task.AsUniTask(),
-                _ => throw new NotImplementedException($"Type {type} was not implemented"),
-            };
-            loader.Dispose();
-
-            bool canceled = false;
-            try
-            {
-                await task;
-            }
-            catch (OperationCanceledException)
-            {
-                canceled = true;
-            }
-            Assert.True(canceled);
-        });
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringUnloadSceneAsync([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc) => UniTask.ToCoroutine(async () =>
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            await LoadFirstScene(loader);
-
-            SceneLoaderType type = GetSceneLoaderType(loader);
-            ILoadSceneInfo sceneInfo = new LoadSceneInfoScene(loader.Manager.GetLastLoadedScene());
-
-            UniTask<Scene> task = default;
-            task = type switch
-            {
-                SceneLoaderType.Async => ((ISceneLoaderAsync)loader).UnloadSceneAsync(sceneInfo).AsTask().AsUniTask(),
-                SceneLoaderType.UniTask => ((ISceneLoaderUniTask)loader).UnloadSceneAsync(sceneInfo),
-                SceneLoaderType.Coroutine => ((ISceneLoaderCoroutine)loader).UnloadSceneAsync(sceneInfo).Task.AsUniTask(),
-                _ => throw new NotImplementedException($"Type {type} was not implemented"),
-            };
-            loader.Dispose();
-
-            bool canceled = false;
-            try
-            {
-                await task;
-            }
-            catch (OperationCanceledException)
-            {
-                canceled = true;
-            }
-            Assert.True(canceled);
-        });
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringUnloadScenesAsync([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes) => UniTask.ToCoroutine(async () =>
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            int sceneCount = targetScenes.Length;
-            loader.LoadScenes(targetScenes);
-
-            await UniTask.WaitUntil(() => loader.Manager.LoadedSceneCount == sceneCount).TimeoutWithoutException(TimeSpan.FromMilliseconds(_defaultTimeout));
-
-            SceneLoaderType type = GetSceneLoaderType(loader);
-
-            UniTask<Scene[]> task = default;
-            task = type switch
-            {
-                SceneLoaderType.Async => ((ISceneLoaderAsync)loader).UnloadScenesAsync(targetScenes).AsTask().AsUniTask(),
-                SceneLoaderType.UniTask => ((ISceneLoaderUniTask)loader).UnloadScenesAsync(targetScenes),
-                SceneLoaderType.Coroutine => ((ISceneLoaderCoroutine)loader).UnloadScenesAsync(targetScenes).Task.AsUniTask(),
-                _ => throw new NotImplementedException($"Type {type} was not implemented"),
-            };
-            loader.Dispose();
-
-            bool canceled = false;
-            try
-            {
-                await task;
-            }
-            catch (OperationCanceledException)
-            {
-                canceled = true;
-            }
-            Assert.True(canceled);
-        });
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringTransitionToSceneAsync([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo targetScene, [ValueSource(nameof(_loadingSceneInfos))] ILoadSceneInfo loadingScene) => UniTask.ToCoroutine(async () =>
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            await LoadFirstScene(loader);
-
-            SceneLoaderType type = GetSceneLoaderType(loader);
-
-            UniTask<Scene> task = default;
-            task = type switch
-            {
-                SceneLoaderType.Async => ((ISceneLoaderAsync)loader).TransitionToSceneAsync(targetScene, loadingScene).AsTask().AsUniTask(),
-                SceneLoaderType.UniTask => ((ISceneLoaderUniTask)loader).TransitionToSceneAsync(targetScene, loadingScene),
-                SceneLoaderType.Coroutine => ((ISceneLoaderCoroutine)loader).TransitionToSceneAsync(targetScene, loadingScene).Task.AsUniTask(),
-                _ => throw new NotImplementedException($"Type {type} was not implemented"),
-            };
-            loader.Dispose();
-
-            bool canceled = false;
-            try
-            {
-                await task;
-            }
-            catch (OperationCanceledException)
-            {
-                canceled = true;
-            }
-            Assert.True(canceled);
-        });
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringTransitionToScenesAsync([ValueSource(nameof(_sceneLoaderCreateFuncs))] Func<ISceneLoader> loaderCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] targetScenes, [ValueSource(nameof(_loadingSceneInfos))] ILoadSceneInfo loadingScene) => UniTask.ToCoroutine(async () =>
-        {
-            ISceneLoader loader = loaderCreateFunc();
-            await LoadFirstScene(loader);
-
-            SceneLoaderType type = GetSceneLoaderType(loader);
-
-            UniTask<Scene[]> task = default;
-            task = type switch
-            {
-                SceneLoaderType.Async => ((ISceneLoaderAsync)loader).TransitionToScenesAsync(targetScenes, 0, loadingScene).AsTask().AsUniTask(),
-                SceneLoaderType.UniTask => ((ISceneLoaderUniTask)loader).TransitionToScenesAsync(targetScenes, 0, loadingScene),
-                SceneLoaderType.Coroutine => ((ISceneLoaderCoroutine)loader).TransitionToScenesAsync(targetScenes, 0, loadingScene).Task.AsUniTask(),
-                _ => throw new NotImplementedException($"Type {type} was not implemented"),
-            };
-            loader.Dispose();
-
-            bool canceled = false;
-            try
-            {
-                await task;
-            }
-            catch (OperationCanceledException)
-            {
-                canceled = true;
-            }
-            Assert.True(canceled);
-        });
-#endif
-
         /// <summary>
         /// Required to test transition scenarios, otherwise the initial (test) scene would be unloaded and stop the tests.
         /// </summary>
-        IEnumerator LoadFirstScene(ISceneLoader sceneLoader)
+        public static IEnumerator LoadFirstScene(ISceneLoader sceneLoader)
         {
             sceneLoader.Manager.SceneLoaded += sceneLoaded;
             sceneLoader.LoadScene(new LoadSceneInfoName(SceneBuilder.SceneNames[1]), true);
@@ -523,7 +230,7 @@ namespace MyGameDevTools.SceneLoading.Tests
 
             var watch = new Stopwatch();
             watch.Start();
-            yield return new WaitUntil(() => loadedScene.IsValid() && loadedScene.isLoaded || watch.ElapsedMilliseconds > _defaultTimeout);
+            yield return new WaitUntil(() => loadedScene.IsValid() && loadedScene.isLoaded || watch.ElapsedMilliseconds > SceneTestEnvironment.DefaultTimeout);
             watch.Stop();
 
             sceneLoader.Manager.SceneLoaded -= sceneLoaded;
@@ -535,20 +242,6 @@ namespace MyGameDevTools.SceneLoading.Tests
             {
                 loadedScene = scene;
             }
-        }
-
-        SceneLoaderType GetSceneLoaderType(ISceneLoader loader)
-        {
-            if (loader is SceneLoaderAsync)
-                return SceneLoaderType.Async;
-#if ENABLE_UNITASK
-            else if (loader is SceneLoaderUniTask)
-                return SceneLoaderType.UniTask;
-#endif
-            else if (loader is SceneLoaderCoroutine)
-                return SceneLoaderType.Coroutine;
-
-            throw new Exception("Unexpected ISceneLoader type");
         }
     }
 }
