+++ Packages/mygamedevtools-scene-loader/Tests/Runtime/SceneManagerTests.cs
@@ -4,7 +4,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using System.Threading;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.TestTools;
@@ -12,14 +11,8 @@ using UnityEngine.TestTools;
 namespace MyGameDevTools.SceneLoading.Tests
 {
     [PrebuildSetup(typeof(SceneTestEnvironment)), PostBuildCleanup(typeof(SceneTestEnvironment))]
-    public class SceneManagerTests
+    public class SceneManagerTests : SceneTestBase
     {
-        // Note: These functions must create new scene managers to correctly test the dispose flow
-        static readonly Func<ISceneManager>[] _sceneManagerCreateFuncs = new Func<ISceneManager>[]
-        {
-            () => new AdvancedSceneManager(),
-        };
-
         static readonly bool[] _setActiveParameterValues = new bool[] { true, false };
 
         static readonly int[] _setIndexActiveParameterValues = new int[] { -1, 1 };
@@ -31,6 +24,8 @@ namespace MyGameDevTools.SceneLoading.Tests
         [OneTimeSetUp]
         public void OneTimeSetup()
         {
+            SceneTestEnvironment.ValidateSceneEnvironment();
+
             ISceneManager[] sceneManagers = SceneTestEnvironment.SceneManagers;
             for (int i = 0; i < sceneManagers.Length; i++)
             {
@@ -60,17 +55,6 @@ namespace MyGameDevTools.SceneLoading.Tests
             _scenesActivated = 0;
             _scenesUnloaded = 0;
             _scenesLoaded = 0;
-        }
-
-        [UnityTearDown]
-        public IEnumerator TearDown()
-        {
-            ISceneManager[] sceneManagers = SceneTestEnvironment.SceneManagers;
-            for (int i = 0; i < sceneManagers.Length; i++)
-                yield return SceneTestUtilities.UnloadManagerScenes(sceneManagers[i]);
-
-            yield return SceneTestUtilities.UnloadRemainingScenes();
-            Assert.AreEqual(1, SceneManager.sceneCount);
         }
 
         [UnityTest]
@@ -334,114 +318,6 @@ namespace MyGameDevTools.SceneLoading.Tests
             yield return new WaitTask<Scene>(task);
 
             Assert.Zero(manager.LoadedSceneCount);
-        }
-
-        [Test]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public void Dispose_Simple([ValueSource(nameof(_sceneManagerCreateFuncs))] Func<ISceneManager> managerCreateFunc)
-        {
-            ISceneManager manager = managerCreateFunc();
-            Assert.DoesNotThrow(manager.Dispose);
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringLoadScene([ValueSource(nameof(_sceneManagerCreateFuncs))] Func<ISceneManager> managerCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo sceneInfo)
-        {
-            ISceneManager manager = managerCreateFunc();
-            WaitTask<Scene> waitTask = new(manager.LoadSceneAsync(sceneInfo).AsTask());
-            manager.Dispose();
-            yield return waitTask;
-            Assert.That(waitTask.Task.IsCompleted);
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dispose_DuringLoadScenes([ValueSource(nameof(_sceneManagerCreateFuncs))] Func<ISceneManager> managerCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] sceneInfos)
-        {
-            ISceneManager manager = managerCreateFunc();
-            WaitTask<Scene[]> waitTask = new(manager.LoadScenesAsync(sceneInfos).AsTask());
-            manager.Dispose();
-            yield return waitTask;
-            Assert.True(waitTask.Task.IsCanceled);
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dipose_DuringUnloadScene([ValueSource(nameof(_sceneManagerCreateFuncs))] Func<ISceneManager> managerCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo sceneInfo)
-        {
-            ISceneManager manager = managerCreateFunc();
-            WaitTask<Scene> waitTask = new(manager.LoadSceneAsync(sceneInfo).AsTask());
-            yield return waitTask;
-
-            waitTask = new(manager.UnloadSceneAsync(sceneInfo).AsTask());
-            manager.Dispose();
-            yield return waitTask;
-            Assert.True(waitTask.Task.IsCanceled);
-        }
-
-        [UnityTest]
-        [Category(SceneTestEnvironment.DisposeCategoryName)]
-        public IEnumerator Dipose_DuringUnloadScenes([ValueSource(nameof(_sceneManagerCreateFuncs))] Func<ISceneManager> managerCreateFunc, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] sceneInfos)
-        {
-            ISceneManager manager = managerCreateFunc();
-            WaitTask<Scene[]> waitTask = new(manager.LoadScenesAsync(sceneInfos).AsTask());
-            yield return waitTask;
-
-            waitTask = new(manager.UnloadScenesAsync(sceneInfos).AsTask());
-            manager.Dispose();
-            yield return waitTask;
-            Assert.True(waitTask.Task.IsCanceled);
-        }
-
-        [UnityTest]
-        public IEnumerator Cancellation_DuringLoadScene([ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SceneManagers))] ISceneManager manager, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo sceneInfo)
-        {
-            CancellationTokenSource tokenSource = new();
-            WaitTask<Scene> waitTask = new(manager.LoadSceneAsync(sceneInfo, token: tokenSource.Token).AsTask());
-            tokenSource.Cancel();
-            yield return waitTask;
-            Assert.That(waitTask.Task.IsCompleted);
-            tokenSource.Dispose();
-        }
-
-        [UnityTest]
-        public IEnumerator Cancellation_DuringLoadScenes([ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SceneManagers))] ISceneManager manager, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] sceneInfos)
-        {
-            CancellationTokenSource tokenSource = new();
-            WaitTask<Scene[]> waitTask = new(manager.LoadScenesAsync(sceneInfos, token: tokenSource.Token).AsTask());
-            tokenSource.Cancel();
-            yield return waitTask;
-            Assert.True(waitTask.Task.IsCanceled);
-            tokenSource.Dispose();
-        }
-
-        [UnityTest]
-        public IEnumerator Cancellation_DuringUnloadScene([ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SceneManagers))] ISceneManager manager, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SingleLoadSceneInfoList))] ILoadSceneInfo sceneInfo)
-        {
-            CancellationTokenSource tokenSource = new();
-            WaitTask<Scene> waitTask = new(manager.LoadSceneAsync(sceneInfo).AsTask());
-            yield return waitTask;
-
-            waitTask = new(manager.UnloadSceneAsync(sceneInfo, token: tokenSource.Token).AsTask());
-            tokenSource.Cancel();
-            yield return waitTask;
-            Assert.True(waitTask.Task.IsCanceled);
-            tokenSource.Dispose();
-        }
-
-        [UnityTest]
-        public IEnumerator Cancellation_DuringUnloadScenes([ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.SceneManagers))] ISceneManager manager, [ValueSource(typeof(SceneTestEnvironment), nameof(SceneTestEnvironment.MultipleLoadSceneInfoList))] ILoadSceneInfo[] sceneInfos)
-        {
-            CancellationTokenSource tokenSource = new();
-            WaitTask<Scene[]> waitTask = new(manager.LoadScenesAsync(sceneInfos).AsTask());
-            yield return waitTask;
-
-            waitTask = new(manager.UnloadScenesAsync(sceneInfos, token: tokenSource.Token).AsTask());
-            tokenSource.Cancel();
-            yield return waitTask;
-            Assert.True(waitTask.Task.IsCanceled);
-            tokenSource.Dispose();
         }
 
         void ReportSceneActivation(Scene previousScene, Scene newScene)
