+++ Packages/mygamedevtools-scene-loader/Tests/Runtime/SceneManager_CancellationTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1873859967aa79a4da9d47b4632097e6
