+++ Packages/mygamedevtools-scene-loader/Tests/Runtime/Utilities/SceneTestBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0584044fe61df264b9fd740bdb67fe4f
